feat: remove interface ITickerData from marketcontext and create independent file for ITickerData
